Move split to module class
