+++ src/split_reduce.cpp
@@ -63,120 +63,6 @@ MIGRAPHX_REGISTER_OP(split_fused_reduce);
 
 static bool is_reduce(const instruction& ins) { return contains(ins.name(), "reduce"); }
 
-struct module_with_inputs
-{
-    module mod;
-    std::vector<instruction_ref> inputs;
-    void replace(instruction_ref ins, instruction_ref rep)
-    {
-        auto it = std::find(inputs.begin(), inputs.end(), ins);
-        if(it == inputs.end())
-            return;
-        *it = rep;
-    }
-    void replace(const std::unordered_map<instruction_ref, instruction_ref>& map_ins)
-    {
-        for(auto& ins : inputs)
-        {
-            if(not contains(map_ins, ins))
-                continue;
-            ins = map_ins.at(ins);
-        }
-    }
-};
-
-static std::vector<instruction_ref>
-select_params(const std::vector<instruction_ref>& instructions,
-              const std::unordered_map<instruction_ref, instruction_ref>& param_map)
-{
-    std::vector<instruction_ref> result;
-    transform_if(
-        instructions.begin(),
-        instructions.end(),
-        std::back_inserter(result),
-        [&](instruction_ref ins) { return contains(param_map, ins); },
-        [&](instruction_ref ins) { return param_map.at(ins); });
-    sort_params(result);
-    return result;
-}
-
-static std::array<module_with_inputs, 2> split_module(module_ref m,
-                                                      const std::vector<instruction_ref>& splits,
-                                                      const std::vector<instruction_ref>& args)
-{
-    std::unordered_map<instruction_ref, instruction_ref> param_map =
-        m->get_ins_param_map(args, true);
-
-    std::unordered_set<instruction_ref> selected_instructions;
-    fix([&](auto self, const std::vector<instruction_ref>& inputs) {
-        for(auto input : inputs)
-        {
-            if(contains(selected_instructions, input))
-                continue;
-            selected_instructions.insert(input);
-            self(input->inputs());
-        }
-    })(splits);
-
-    std::vector<instruction_ref> instructions1;
-    // TODO: copy_if
-    for(auto ins : iterator_for(*m))
-    {
-        if(not contains(selected_instructions, ins))
-            continue;
-        instructions1.push_back(ins);
-    }
-
-    std::vector<instruction_ref> inputs1 = select_params(instructions1, param_map);
-    module m1;
-    std::unordered_map<instruction_ref, instruction_ref> map_ins1;
-    m1.add_instructions(instructions1, &map_ins1);
-    std::vector<instruction_ref> outputs;
-    std::transform(splits.begin(),
-                   splits.end(),
-                   std::back_inserter(outputs),
-                   [&](instruction_ref ins) { return map_ins1.at(ins); });
-    m1.add_return(outputs);
-
-    std::vector<instruction_ref> instructions2;
-    for(auto ins : iterator_for(*m))
-    {
-        if(contains(selected_instructions, ins))
-            continue;
-        // Input params can be used in both modules
-        std::vector<instruction_ref> input_params;
-        std::copy_if(ins->inputs().begin(),
-                     ins->inputs().end(),
-                     std::back_inserter(input_params),
-                     [&](instruction_ref input) {
-                         if(input->name() != "@param")
-                             return false;
-                         return not contains(instructions2, input);
-                     });
-        instructions2.insert(instructions2.end(), input_params.begin(), input_params.end());
-        instructions2.push_back(ins);
-    }
-
-    std::vector<instruction_ref> inputs2 = select_params(instructions2, param_map);
-    inputs2.insert(inputs2.begin(), splits.begin(), splits.end());
-    module m2;
-    std::unordered_map<instruction_ref, instruction_ref> map_ins2;
-    std::size_t n = 0;
-    for(auto ins : splits)
-        map_ins2[ins] = m2.add_parameter(param_name(n++), ins->get_shape().as_standard());
-    for(auto ins : iterator_for(*m))
-    {
-        if(ins->name() != "@param")
-            continue;
-        if(not contains(instructions2, ins))
-            continue;
-        map_ins2[ins] = m2.add_parameter(param_name(n++), ins->get_shape().as_standard());
-    }
-    auto r = m2.add_instructions(instructions2, &map_ins2);
-    m2.add_return(r);
-    return {{{std::move(m1), std::move(inputs1)}, {std::move(m2), std::move(inputs2)}}};
-}
-
 static std::vector<instruction_ref> find_split(module_ref rm)
 {
     std::vector<instruction_ref> result;
@@ -221,7 +107,7 @@ void split_reduce::apply(module_pass_manager& mpm) const
         auto axes = v["axes"].to_vector<std::int64_t>();
         // TODO: Check reduction size
 
-        auto mp  = split_module(rm, splits, ins->inputs());
+        auto mp  =  rm->split(ins->inputs(), splits);
         auto* m1 = mpm.create_module(rm->name() + "_0", std::move(mp[0].mod));
         auto* m2 = mpm.create_module(rm->name() + "_1", std::move(mp[1].mod));
         m1->set_bypass();
